* Add player fade-in to settings.cfg creation, edit and save * Start and Stop server buttons now a single dual function button * Settings and logging (settings.cfg) now in it's own class and dialog * General UI overhaul * UI fixes * Change to settings.cfg read logic
