+++ Projects/CoX/Utilities/SEGSAdmin/GetIPDialog.ui
@@ -9,8 +9,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>282</width>
-    <height>90</height>
+    <width>294</width>
+    <height>97</height>
    </rect>
   </property>
   <property name="windowTitle">
