+++ Projects/CoX/Utilities/SEGSAdmin/resources.qrc
@@ -1,11 +1,15 @@
 <RCC>
     <qresource prefix="/logo">
         <file>logo3.png</file>
+        <file>logo4.png</file>
     </qresource>
     <qresource prefix="/icons">
         <file>icon_warning.png</file>
         <file>icon_good.png</file>
         <file>segs.bmp</file>
         <file>icon_lookup.png</file>
+    </qresource>
+    <qresource prefix="/fonts">
+        <file>dejavusanscondensed.ttf</file>
     </qresource>
 </RCC>
